+++ lib/file/split-markdown/core/splitter.js
