+++ lib/file/text-splitter.js
@@ -7,7 +7,7 @@ import { getProject } from '@/lib/db/projects';
 import { getChunkByProjectId, saveChunks } from '@/lib/db/chunks';
 
 // 导入Markdown分割工具
-const markdownSplitter = require('./split-mardown/index');
+const markdownSplitter = require('./split-markdown/index');
 
 /**
  * 分割项目中的Markdown文件
